third_party/blink/web_tests/fast/css/image-set-parsing-invalid-expected.txt

Test the parsing of the image-set functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Too many url parameters : image-set(url(#a #b)
PASS cssRule is ""


Too many url parameters : -webkit-image-set(url(#a #b)
PASS cssRule is ""


No x : image-set(url('#a') 1
PASS cssRule is ""


No x : -webkit-image-set(url('#a') 1
PASS cssRule is ""


No comma : image-set(url('#a') 1x url('#b') 2x
PASS cssRule is ""


No comma : -webkit-image-set(url('#a') 1x url('#b') 2x
PASS cssRule is ""


Too many scale factor parameters : image-set(url('#a') 1x 2x
PASS cssRule is ""


Too many scale factor parameters : -webkit-image-set(url('#a') 1x 2x
PASS cssRule is ""
PASS successfullyParsed is true

TEST COMPLETE
